+++ NEWS.md
@@ -6,7 +6,7 @@
 * Checks the value of the `date1904` attribute for `"false"` or `"1"` to support files
   created by the `openxlsx` package (#8).
 * Fixed a bug that only imported the first line of multiline comments (#9).
-* Encodes comment text as UTF-8 (#12).
+* Encodes cell and comment text as UTF-8 (#12).
 * Finds worksheets more reliably in files not created by Excel (part of #13).
 * Imports dates with greater precision (part of #14).